#128: Restructure layout of OAV and Snapshot
