+++ src/artemis/devices/oav/__init__.py
@@ -1,0 +1,22 @@
+from ophyd import ADComponent as ADC
+from ophyd import (
+    AreaDetector,
+    CamBase,
+    Component,
+    HDF5Plugin,
+    OverlayPlugin,
+    ProcessPlugin,
+    ROIPlugin,
+)
+
+from src.artemis.devices.oav.snapshot import Snapshot
+
+
+class OAV(AreaDetector):
+    cam = ADC(CamBase, "CAM:")
+    roi = ADC(ROIPlugin, "ROI:")
+    proc = ADC(ProcessPlugin, "PROC:")
+    over = ADC(OverlayPlugin, "OVER:")
+    tiff = ADC(OverlayPlugin, "TIFF:")
+    hdf5 = ADC(HDF5Plugin, "HDF5:")
+    snapshot: Snapshot = Component(Snapshot, ":MJPG")
