+++ src/artemis/devices/oav/snapshot.py
@@ -2,20 +2,7 @@ import threading
 from pathlib import Path
 
 import requests
-from ophyd import ADComponent as ADC
-from ophyd import (
-    AreaDetector,
-    CamBase,
-    Component,
-    Device,
-    DeviceStatus,
-    EpicsSignal,
-    HDF5Plugin,
-    OverlayPlugin,
-    ProcessPlugin,
-    ROIPlugin,
-    Signal,
-)
+from ophyd import Component, Device, DeviceStatus, EpicsSignal, Signal
 from PIL import Image
 
 
@@ -49,13 +36,3 @@ class Snapshot(Device):
 
     def post_processing(self, image: Image.Image):
         pass
-
-
-class OAV(AreaDetector):
-    cam = ADC(CamBase, "CAM:")
-    roi = ADC(ROIPlugin, "ROI:")
-    proc = ADC(ProcessPlugin, "PROC:")
-    over = ADC(OverlayPlugin, "OVER:")
-    tiff = ADC(OverlayPlugin, "TIFF:")
-    hdf5 = ADC(HDF5Plugin, "HDF5:")
-    snapshot: Snapshot = Component(Snapshot, ":MJPG")
